Refactor publishstatus logic into own class
