+++ PublishModal.ts
@@ -2,25 +2,23 @@ import DigitalGardenSettings from "DigitalGardenSettings";
 import { IDigitalGardenSiteManager } from "DigitalGardenSiteManager";
 import { App, Modal, TFile } from "obsidian";
 import { IPublisher } from "Publisher";
+import { IPublishStatusManager } from "PublishStatusManager";
 import { generateBlobHash } from "utils";
 
 export class PublishModal {
     modal: Modal;
-    siteManager: IDigitalGardenSiteManager;
     settings: DigitalGardenSettings;
-    publisher: IPublisher;
+    publishStatusManager: IPublishStatusManager;
 
     publishedContainer: HTMLElement;
     changedContainer: HTMLElement;
     deletedContainer: HTMLElement;
     unpublishedContainer: HTMLElement;
 
-    constructor(app: App, siteManager: IDigitalGardenSiteManager, publisher: IPublisher,
-        settings: DigitalGardenSettings) {
+    constructor(app: App, publishStatusManager: IPublishStatusManager, settings: DigitalGardenSettings) {
         this.modal = new Modal(app)
-        this.siteManager = siteManager;
         this.settings = settings;
-        this.publisher = publisher;
+        this.publishStatusManager = publishStatusManager;
 
         this.initialize();
     }
@@ -77,60 +75,14 @@ export class PublishModal {
         }
     }
     async populateWithNotes() {
-        const publishStatus = await this.buildPublishStatus();
+        const publishStatus = await this.publishStatusManager.getPublishStatus();
         publishStatus.publishedNotes.map(file => this.publishedContainer.createEl("li", { text: file.path }));
         publishStatus.unpublishedNotes.map(file => this.unpublishedContainer.createEl("li", { text: file.path }));
         publishStatus.changedNotes.map(file => this.changedContainer.createEl("li", { text: file.path }));
         publishStatus.deletedNotePaths.map(path => this.deletedContainer.createEl("li", { text: path }));
     }
 
-    async buildPublishStatus(): Promise<PublishStatus> {
-        const unpublishedNotes: Array<TFile> = [];
-        const publishedNotes: Array<TFile> = [];
-        const changedNotes: Array<TFile> = [];
-
-        const deletedNotePaths: Array<string> = [];
-
-        const remoteNoteHashes = await this.siteManager.getNoteHashes();
-        const marked = await this.publisher.getFilesMarkedForPublishing();
-
-        for (const file of marked) {
-            const content = await this.publisher.generateMarkdown(file);
-
-            const localHash = generateBlobHash(content);
-            const remoteHash = remoteNoteHashes[file.path];
-            if (!remoteHash) {
-                unpublishedNotes.push(file);
-            }
-            else if (remoteHash === localHash) {
-                publishedNotes.push(file);
-            }
-            else {
-                changedNotes.push(file);
-            }
-        }
-
-        Object.keys(remoteNoteHashes).forEach(key => {
-            if (!marked.find(f => f.path === key)) {
-                deletedNotePaths.push(key);
-            }
-        });
-
-        unpublishedNotes.sort((a, b) => a.path > b.path ? 1 : -1);
-        publishedNotes.sort((a, b) => a.path > b.path ? 1 : -1);
-        changedNotes.sort((a, b) => a.path > b.path ? 1 : -1);
-        deletedNotePaths.sort((a, b) => a > b ? 1 : -1);
-        return { unpublishedNotes, publishedNotes, changedNotes, deletedNotePaths };
-    }
-
     open() {
         this.modal.open();
     }
-}
-
-interface PublishStatus{
-    unpublishedNotes: Array<TFile>;
-    publishedNotes: Array<TFile>;
-    changedNotes: Array<TFile>;
-    deletedNotePaths: Array<string>;
 }
